refactor(core): extract agent ratings and reviews into agent-rating
apps/core/src/helpers/agent.ts was 762 lines, over the 750-line hard ceiling
in CLAUDE.md. Split by responsibility, following the precedent this PR set
with agent-cost: user ratings and reviews (public aggregates, the review
feed, and the caller's own review) move to helpers/agent-rating.ts, leaving
agent.ts with agent identity, availability, rail readiness and execution
metrics.

agent.ts is now 561 lines and agent-rating.ts 220. Tests move with the code,
and getAgentRatingDistribution — extracted untested — gains coverage for the
zero-filled buckets and the isHidden filter. Import sites and the
vi.mock("@/helpers/agent") factories in the agent route tests are updated to
the new module. No behaviour change.
