+++ apps/core/src/routes/v1/agents/[id]/ratings/post.test.ts
@@ -25,6 +25,9 @@ const {
 
 vi.mock("@/helpers/agent", () => ({
   requireAvailableAgentOrThrow: requireAvailableAgentOrThrowMock,
+}));
+
+vi.mock("@/helpers/agent-rating", () => ({
   upsertUserAgentReview: upsertUserAgentReviewMock,
 }));
