+++ apps/core/src/routes/v1/agents/get.test.ts
@@ -40,7 +40,6 @@ vi.mock("@/helpers/agent", () => ({
   AGENT_PRICING_READ_TRANSACTION_OPTIONS: { isolationLevel: "RepeatableRead" },
   getCardanoV2ReadySources: () => Promise.resolve([]),
   buildAvailableAgentWhereClause: buildAvailableAgentWhereClauseMock,
-  calculateAgentRatings: calculateAgentRatingsMock,
   calculateAverageExecutionTimes: calculateAverageExecutionTimesMock,
   getAgentAuthorImage: getAgentAuthorImageMock,
   getAgentDescription: getAgentDescriptionMock,
@@ -48,6 +47,10 @@ vi.mock("@/helpers/agent", () => ({
   getAgentImage: getAgentImageMock,
   getAgentName: getAgentNameMock,
   getCreditCostsOrThrow: getCreditCostsOrThrowMock,
+}));
+
+vi.mock("@/helpers/agent-rating", () => ({
+  calculateAgentRatings: calculateAgentRatingsMock,
 }));
 
 vi.mock("@/helpers/agent-cost", () => ({
